// src/entities/working-shifts/model/types.ts
export type typeWorkingShifts = {
    id: string
    createdAt: string
    createdBy: string
    merchantId: string
    openedAt: string
    closedAt: string
    storeId: string
    storeName: string
    terminalId: string
    terminalSerialNumber: string
    cashierId: string
    cashierName: string
    openedBy: string
    closedBy: string
    totalCashIncome: number
    totalCardIncome: number
    totalQRIncome: number
    totalTransferIncome: number
    totalOtherIncome: number
    salesCount: number
    refundsCount: number
    totalCashRefunds: number
    totalCardRefunds: number
    totalQRRefunds: number
    totalTransferRefunds: number
    totalOtherRefunds: number
}